Move to preprod environment for some tests

+++ src/main/kotlin/io/newm/kogmios/protocols/model/PlutusV1.kt
@@ -504,5 +504,5 @@ data class PlutusV1(
     val verifyEd25519SignatureCpuArgumentsSlope: BigInteger,
     @Contextual
     @SerialName("verifyEd25519Signature-memory-arguments")
-    val verifyEd25519SignatureMemoryArguments: Int
+    val verifyEd25519SignatureMemoryArguments: BigInteger,
 )
